Struct project1::page_table::PageTable

pub struct PageTable(_);

4-level page table of x86_64.

Implementations

impl PageTable

pub fn new() -> Self

Create a empty page table.

pub fn pa(&self) -> Pa

Get physical address of this page table.

pub fn map(
&mut self,
va: Va,
pg: Page,
perm: Permission
) -> Result<(), PageTableMappingError>

Map pg into va with permission perm.

pub unsafe fn do_map(
&mut self,
va: Va,
pa: Pa,
perm: Permission
) -> Result<(), PageTableMappingError>

Map pa into va with permission perm.

Safety

pa must be valid.

pub fn unmap(&mut self, va: Va) -> Result<Page, PageTableMappingError>

Unmap the va and returns Page that was mapped to va.

pub fn walk(&self, va: Va) -> Result<&Pte, PageTableMappingError>

Walk the page table and return corresponding pte of the va if exist.